requirements.txt
transformers>=4.56.0
accelerate
pillow
torchvision
opencv-python-headless
huggingface_hub
psutil

# GGUF Node Requirements (optional)
# pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
